polish trace moving
