+++ paddle/pten/kernels/gpu/trace_kernel.cu
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "paddle/pten/kernels/trace_kernel.h"
+
 #include "paddle/pten/backends/gpu/gpu_context.h"
 #include "paddle/pten/core/kernel_registry.h"
+#include "paddle/pten/kernels/funcs/diagonal.h"
 #include "paddle/pten/kernels/gpu/reduce.h"
-#include "paddle/pten/kernels/impl/trace_kernel_impl.h"
-#include "paddle/pten/kernels/trace_kernel.h"
 
 namespace pten {
 
@@ -27,8 +28,8 @@ void TraceKernel(const Context& ctx,
                  int axis1,
                  int axis2,
                  DenseTensor* out) {
-  T* out_data = out->mutable_data<T>(ctx.GetPlace());
-  auto diag = Diagonal<T, Context>(ctx, &x, offset, axis1, axis2);
+  T* out_data = ctx.template Alloc<T>(out);
+  auto diag = funcs::Diagonal<T, Context>(ctx, &x, offset, axis1, axis2);
   if (diag.numel() > 0) {
     auto stream = ctx.stream();
     std::vector<int> reduce_dims;
